% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{bomonth}
\alias{bomonth}
\title{Converts to Beginning of Month}
\usage{
bomonth(date, n_months = 0)
}
\arguments{
\item{date}{A date value}

\item{n_months}{Number of months from current date}
}
\value{
The beginning month of the input date
}
\description{
Converts to Beginning of Month
}
\examples{
bomonth(as.Date("2024-02-12"))
}
