% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{date_yyyymm}
\alias{date_yyyymm}
\title{A function that convert a date into a numeric yyyymm form}
\usage{
date_yyyymm(date)
}
\arguments{
\item{date}{A date value}
}
\value{
A numeric representation of a date in the for yyyymm
}
\description{
A function that convert a date into a numeric yyyymm form
}
\examples{
date_yyyymm(as.Date('2023-03-23'))
}
