% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{edate}
\alias{edate}
\title{Add months to date}
\usage{
edate(date, months)
}
\arguments{
\item{date}{The date value}

\item{months}{The number of months to add}
}
\value{
A date value "months" from the input date
}
\description{
Add months to date
}
\examples{
bomonth(as.Date("2024-02-12"),4)
}
