% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{eomonth}
\alias{eomonth}
\title{Converts to End of Month}
\usage{
eomonth(date)
}
\arguments{
\item{date}{A date value}
}
\value{
A date representing the end of the month of the input date
}
\description{
Converts to End of Month
}
\examples{
eomonth(as.Date("2024-02-12"))
}
